[ie/youtube] Split into package (#12557)
Authored by: coletdjnz
